+++ features/step-definitions/clj-refactor-steps.el
@@ -64,10 +64,6 @@
        (lambda ()
          (setq cljr-use-multiple-cursors nil)))
 
-(Given "^I enable cljr-magic-uses"
-       (lambda ()
-         (setq cljr-magic-uses t)))
-
 (Given "^I add \"\\([^\"]+\\)\" to cljr-magic-use-namespaces$"
        (lambda (namespace)
          (add-to-list 'cljr-magic-use-namespaces namespace)))